src/main/java/falcon/mvc/recipes/converters/NotesToNotesCommand.java
package falcon.mvc.recipes.converters;

import falcon.mvc.recipes.commands.NotesCommand;
import falcon.mvc.recipes.domains.Notes;
import lombok.Synchronized;
import org.springframework.core.convert.converter.Converter;
import org.springframework.lang.Nullable;
import org.springframework.stereotype.Component;

@Component
public class NotesToNotesCommand implements Converter<Notes, NotesCommand> {

    @Synchronized
    @Nullable
    @Override
    public NotesCommand convert(Notes source) {
        if (source == null) {
            throw new NullPointerException("Source object is null");
        }

        final NotesCommand notes = new NotesCommand();
        notes.setId(source.getId());
        notes.setRecipeNotes(source.getRecipeNotes());
        return notes;
    }
}
